+++ BlinkCard/android/src/main/java/com/microblink/blinkcard/capacitor/MicroblinkPlugin.java
@@ -12,6 +12,7 @@ import com.microblink.blinkcard.intent.IntentDataTransferMode;
 import com.microblink.blinkcard.uisettings.UISettings;
 import com.microblink.blinkcard.capacitor.overlays.OverlaySettingsSerializers;
 import com.microblink.blinkcard.capacitor.recognizers.RecognizerSerializers;
+import com.microblink.blinkcard.locale.LanguageUtils;
 import com.getcapacitor.annotation.ActivityCallback;
 import com.getcapacitor.annotation.CapacitorPlugin;
 import com.getcapacitor.JSObject;
@@ -40,6 +41,7 @@ public class MicroblinkPlugin extends Plugin {
         JSObject jsLicenses = call.getObject("license");
 
         setLicense(jsLicenses);
+        setLanguage(jsOverlaySettings);
 
         recognizerBundle = RecognizerSerializers.INSTANCE.deserializeRecognizerCollection(jsRecognizerCollection);
         UISettings uiSettings = OverlaySettingsSerializers.INSTANCE.getOverlaySettings(getContext(), jsOverlaySettings, recognizerBundle);
@@ -84,6 +86,14 @@ public class MicroblinkPlugin extends Plugin {
             MicroblinkSDK.setLicenseKey(androidLicense, licensee, context);
         }
         MicroblinkSDK.setIntentDataTransferMode(IntentDataTransferMode.PERSISTED_OPTIMISED);
+    }
+
+    private void setLanguage(JSONObject jsonOverlaySettings) {
+        try {
+            LanguageUtils.setLanguageAndCountry(jsonOverlaySettings.getString("language"),
+                    jsonOverlaySettings.getString("country"),
+                    getContext());
+        } catch (Exception e) {}
     }
 
     private void startActivityForResult(PluginCall call, String callbackMethod, UISettings settings) {