+++ BlinkCard/ios/Plugin/MBCBlinkCardSerializationUtils.h
@@ -10,7 +10,7 @@
 #import <Foundation/Foundation.h>
 
 @interface MBCBlinkCardSerializationUtils : NSObject
-
++(NSDictionary *) serializeDocumentLivenessCheckResult:(MBCDocumentLivenessCheckResult * _Nullable)documentlivenessCheckResult;
 +(MBCCardNumberAnonymizationSettings *) deserializeMBCardNumberAnonymizationSettings:(NSDictionary * _Nullable)jsonCardNumberAnonymizationSettings;
 +(MBCBlinkCardAnonymizationSettings *) deserializeMBBlinkCardAnonymizationSettings:(NSDictionary * _Nullable)jsonBlinkCardAnonymizationSettings;