+++ BlinkCard/ios/Plugin/MicroblinkModule/Recognizers/Wrappers/MBCLegacyBlinkCardEliteRecognizerWrapper.h
@@ -1,6 +1,0 @@
-#import "MBCRecognizerWrapper.h"
-#import <BlinkCard/BlinkCard.h>
-
-@interface MBCLegacyBlinkCardEliteRecognizerCreator : NSObject<MBCRecognizerCreator>
-
-@end