+++ BlinkCard/ios/Plugin/MicroblinkModule/Recognizers/Wrappers/MBCLegacyBlinkCardRecognizerWrapper.h
@@ -1,6 +1,0 @@
-#import "MBCRecognizerWrapper.h"
-#import <BlinkCard/BlinkCard.h>
-
-@interface MBCLegacyBlinkCardRecognizerCreator : NSObject<MBCRecognizerCreator>
-
-@end