+++ BlinkCard/src/recognizers/blinkCardRecognizer.ts
@@ -4,15 +4,15 @@ import {
     Point,
     Quadrilateral,
     
-    LegacyCardIssuer,
     Issuer,
     BlinkCardProcessingStatus,
     BlinkCardAnonymizationMode,
     CardNumberAnonymizationSettings,
     BlinkCardAnonymizationSettings,
+    DocumentLivenessCheckResult,
+    BlinkCardMatchLevel,
     
     ImageExtensionFactors,
-    DataMatchResult,
 } from '../types'
 
 /* tslint:disable:no-unused-variable */
@@ -26,73 +26,88 @@ export class BlinkCardRecognizerResult extends RecognizerResult {
     /**
      * The payment card number.
      */
-    cardNumber: string;
+    cardNumber?: string;
     
     /**
      * The payment card number prefix.
      */
-    cardNumberPrefix: string;
+    cardNumberPrefix?: string;
     
     /**
      * The payment card number is valid
      */
-    cardNumberValid: boolean;
+    cardNumberValid?: boolean;
     
     /**
      *  Payment card's security code/value.
      */
-    cvv: string;
+    cvv?: string;
+    
+    /**
+     * Document liveness check (screen, photocopy, hand presence) which can pass or fail.
+     */
+    documentLivenessCheck?: DocumentLivenessCheckResult;
     
     /**
      * The payment card's expiry date.
      */
-    expiryDate: Date;
+    expiryDate?: Date;
     
     /**
-     * Wheater the first scanned side is blurred.
+     * Whether the first scanned side is anonymized.
      */
-    firstSideBlurred: boolean;
+    firstSideAnonymized?: boolean;
+    
+    /**
+     * Whether the first scanned side is blurred.
+     */
+    firstSideBlurred?: boolean;
     
     /**
      * Full image of the payment card from first side recognition.
      */
-    firstSideFullDocumentImage: string;
+    firstSideFullDocumentImage?: string;
     
     /**
      * Payment card's IBAN.
      */
-    iban: string;
+    iban?: string;
     
     /**
      * Payment card's issuing network.
      */
-    issuer: Issuer;
+    issuer?: Issuer;
     
     /**
      * Information about the payment card owner (name, company, etc.).
      */
-    owner: string;
+    owner?: string;
     
     /**
      * Status of the last recognition process.
      */
-    processingStatus: BlinkCardProcessingStatus;
+    processingStatus?: BlinkCardProcessingStatus;
     
     /**
      * Returns true if recognizer has finished scanning first side and is now scanning back side,
          * false if it's still scanning first side.
      */
-    scanningFirstSideDone: boolean;
+    scanningFirstSideDone?: boolean;
     
     /**
-     * Wheater the second scanned side is blurred.
+     * Whether the second scanned side is anonymized.
      */
-    secondSideBlurred: boolean;
+    secondSideAnonymized?: boolean;
+    
+    /**
+     * Whether the second scanned side is blurred.
+     */
+    secondSideBlurred?: boolean;
     
     /**
      * Full image of the payment card from second side recognition.
      */
-    secondSideFullDocumentImage: string;
+    secondSideFullDocumentImage?: string;
     
 
     constructor(nativeResult: any) {
@@ -119,12 +134,22 @@ export class BlinkCardRecognizerResult extends RecognizerResult {
         this.cvv = nativeResult.cvv;
         
         /**
+         * Document liveness check (screen, photocopy, hand presence) which can pass or fail.
+         */
+        this.documentLivenessCheck = nativeResult.documentLivenessCheck;
+        
+        /**
          * The payment card's expiry date.
          */
         this.expiryDate = nativeResult.expiryDate != null ? new Date(nativeResult.expiryDate) : null;
         
         /**
-         * Wheater the first scanned side is blurred.
+         * Whether the first scanned side is anonymized.
+         */
+        this.firstSideAnonymized = nativeResult.firstSideAnonymized;
+        
+        /**
+         * Whether the first scanned side is blurred.
          */
         this.firstSideBlurred = nativeResult.firstSideBlurred;
         
@@ -160,7 +185,12 @@ export class BlinkCardRecognizerResult extends RecognizerResult {
         this.scanningFirstSideDone = nativeResult.scanningFirstSideDone;
         
         /**
-         * Wheater the second scanned side is blurred.
+         * Whether the second scanned side is anonymized.
+         */
+        this.secondSideAnonymized = nativeResult.secondSideAnonymized;
+        
+        /**
+         * Whether the second scanned side is blurred.
          */
         this.secondSideBlurred = nativeResult.secondSideBlurred;
         
@@ -184,6 +214,13 @@ export class BlinkCardRecognizer extends Recognizer {
          * 
      */
     allowBlurFilter: boolean;
+    
+    /**
+     * Whether invalid card number is accepted.
+         * 
+         * 
+     */
+    allowInvalidCardNumber: boolean;
     
     /**
      * Defines whether sensitive data should be redacted from the result.
@@ -237,6 +274,20 @@ export class BlinkCardRecognizer extends Recognizer {
     fullDocumentImageExtensionFactors: ImageExtensionFactors;
     
     /**
+     * This parameter is used to adjust heuristics that eliminate cases when the hand is present.
+         * 
+         * 
+     */
+    handDocumentOverlapThreshold: number;
+    
+    /**
+     * Hand scale is calculated as a ratio between area of hand mask and document mask.
+         * 
+         * 
+     */
+    handScaleThreshold: number;
+    
+    /**
      * Pading is a minimum distance from the edge of the frame and is defined as a percentage of the frame width. Default value is 0.0f and in that case
          * padding edge and image edge are the same.
          * Recommended value is 0.02f.
@@ -246,11 +297,25 @@ export class BlinkCardRecognizer extends Recognizer {
     paddingEdge: number;
     
     /**
+     * Photocopy analysis match level - higher if stricter.
+         * 
+         * 
+     */
+    photocopyAnalysisMatchLevel: BlinkCardMatchLevel;
+    
+    /**
      * Sets whether full document image of ID card should be extracted.
          * 
          * 
      */
     returnFullDocumentImage: boolean;
+    
+    /**
+     * Screen analysis match level - higher if stricter.
+         * 
+         * 
+     */
+    screenAnalysisMatchLevel: BlinkCardMatchLevel;
     
 
     constructor() {
@@ -262,6 +327,13 @@ export class BlinkCardRecognizer extends Recognizer {
          * 
          */
         this.allowBlurFilter = true;
+        
+        /**
+         * Whether invalid card number is accepted.
+         * 
+         * 
+         */
+        this.allowInvalidCardNumber = false;
         
         /**
          * Defines whether sensitive data should be redacted from the result.
@@ -315,6 +387,20 @@ export class BlinkCardRecognizer extends Recognizer {
         this.fullDocumentImageExtensionFactors = new ImageExtensionFactors();
         
         /**
+         * This parameter is used to adjust heuristics that eliminate cases when the hand is present.
+         * 
+         * 
+         */
+        this.handDocumentOverlapThreshold = 0.05;
+        
+        /**
+         * Hand scale is calculated as a ratio between area of hand mask and document mask.
+         * 
+         * 
+         */
+        this.handScaleThreshold = 0.15;
+        
+        /**
          * Pading is a minimum distance from the edge of the frame and is defined as a percentage of the frame width. Default value is 0.0f and in that case
          * padding edge and image edge are the same.
          * Recommended value is 0.02f.
@@ -324,11 +410,25 @@ export class BlinkCardRecognizer extends Recognizer {
         this.paddingEdge = 0.0;
         
         /**
+         * Photocopy analysis match level - higher if stricter.
+         * 
+         * 
+         */
+        this.photocopyAnalysisMatchLevel = BlinkCardMatchLevel.Level5;
+        
+        /**
          * Sets whether full document image of ID card should be extracted.
          * 
          * 
          */
         this.returnFullDocumentImage = false;
+        
+        /**
+         * Screen analysis match level - higher if stricter.
+         * 
+         * 
+         */
+        this.screenAnalysisMatchLevel = BlinkCardMatchLevel.Level5;
         
 
 	this.createResultFromNative = (nativeResult: any) => { return new BlinkCardRecognizerResult(nativeResult); };